+++ NEWS.md
@@ -1,5 +1,10 @@
 # eDITH 0.1.2.9000
 
+## Major changes
+
+- `run_eDITH_optim`: log-posterior is now maximized. Default arguments for prior distributions
+have been added.
+
 ## Minor changes
 
 - `run_eDITH_BT`, `run_eDITH_optim`: updated error messages when unsuitable input is used.